Use modified DeSmuME 0.9.11 core from iNDS-Team
https://github.com/iNDS-Team/desmume/tree/7138e75ab809bbd8a306259cb7075a132f583c9b

+++ app/src/main/jni/desmume/src/gfx3d.cpp
@@ -1770,7 +1770,7 @@ void gfx3d_glAlphaFunc(u32 v)
 
 unsigned int gfx3d_glGetPosRes(unsigned int index)
 {
-	return (unsigned int)(PTcoords[index] * 4096.0f);
+	return (unsigned int)(int)(PTcoords[index] * 4096.0f);
 }
 
 //#define _3D_LOG_EXEC
